services/updater/requirements.txt
fastapi==0.115.8
typing-inspect==0.9.0
fastapi-utils==0.8.0
python-dateutil==2.9.0.post0
httpx==0.28.1
uvicorn==0.34.0